user-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlmodel==0.0.16
httpx==0.25.2
python-jose==3.3.0
email-validator==2.2.0
